The Big Optimization
- Individual faces are now exported instead of entire blocks, increasing
exporting time, but creating a scene file that takes FAR less time to
parse.  Exporting will be sped up in a later commit with multithreading.
- Sample region has been expanded.
- A lot of code was added, but is currently commented out and will
probably be removed in the next commit.

minecraft.pov template changes:
- Blocks now have a higher diffuse, making them brighter.
- Some macros were added that were used during development of The Big
Optimization, but are no longer used and will probably be removed in the
next commit, as well as some other macros.
- The water texture has been changed to a non-animated version that
looks nicer.
- Camera FOV angle has been increased.
- Gray threshold increased to reduce the blue tint from the sky.
- Sun position has changed to create a greater emphasis on shadows.